+++ inst/paramtest/test_paramtest_classif.glmnet.R
@@ -14,7 +14,8 @@ test_that("classif.glmnet", {
     "type.measure", # only used by cv.glmnet
     "family", # handled by mlr3
     "itrace", # supported via param trace.it
-    "factory" # only used in scripts, no effect within mlr3
+    "factory", # only used in scripts, no effect within mlr3
+    "offset" # handled by mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -32,7 +33,9 @@ test_that("predict classif.glmnet", {
   exclude = c(
     "object", # handled via mlr3
     "newx", # handled via mlr3
-    "type" # handled via mlr3
+    "type", # handled via mlr3
+    "newoffset", # handled via mlr3
+    "use_pred_offset" # handled via mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "predict")