+++ inst/paramtest/test_paramtest_regr.cv_glmnet.R
@@ -10,7 +10,8 @@ test_that("regr.cv_glmnet", {
     "y", # handled by mlr3
     "weights", # handled by mlr3
     "itrace", # supported via param trace.it
-    "factory" # only used in scripts, no effect within mlr3
+    "factory", # only used in scripts, no effect within mlr3
+    "offset" # handled by mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -28,7 +29,8 @@ test_that("predict regr.cv_glmnet", {
   exclude = c(
     "object", # handled via mlr3
     "newx", # handled via mlr3
-    "predict.gamma" # renamed from gamma
+    "predict.gamma", # renamed from gamma
+    "use_pred_offset" # handled by mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "predict")