+++ inst/paramtest/test_paramtest_regr.glmnet.R
@@ -11,7 +11,8 @@ test_that("regr.glmnet", {
     "foldid", # not used by learner
     "type.measure", # only used by cv.glmnet
     "itrace", # supported via param trace.it
-    "factory" # only used in scripts, no effect within mlr3
+    "factory", # only used in scripts, no effect within mlr3
+    "offset" # handled by mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -29,7 +30,9 @@ test_that("predict regr.glmnet", {
   exclude = c(
     "object", # handled via mlr3
     "newx", # handled via mlr3
-    "type" # handled via mlr3
+    "type", # handled via mlr3
+    "newoffset", # handled via mlr3
+    "use_pred_offset" # handled via mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "predict")