+++ inst/paramtest/test_paramtest_regr.lm.R
@@ -10,7 +10,8 @@ test_that("regr.lm", {
     "na.action", # handled via mlr3
     "method", # handled via mlr3
     "subset", # handled via mlr3
-    "contrasts" # handled via mlr3
+    "contrasts", # handled via mlr3
+    "offset" # handled via mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -32,7 +33,8 @@ test_that("predict regr.lm", {
     "na.action", # handled via mlr3
     "terms", # not supported by mlr3 learner
     "weights", # handled via mlr3
-    "se.fit" # controlled via predict type
+    "se.fit", # controlled via predict type
+    "use_pred_offset" # handled via mlr3
   )
 
   ParamTest = run_paramtest(learner, fun, exclude, tag = "predict")